example/requirements.txt
pyspark-event-correlation 1.0.0
